webapp/requirements.txt
streamlit>=1.30.0
PyGithub>=2.1.0
pyyaml>=6.0
pandas>=2.0.0
requests>=2.28.0
